backend/python-fastapi/pyproject.toml
[tool.poetry]
name = "social-media-app"
version = "0.1.0"
description = ""
authors = ["danielbom <daniel.rodrigues.45@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.79.0"}
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.39"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
blue = "^0.9.0"
flake8 = "^4.0.1"
mypy = "^0.971"
bandit = "^1.7.4"
pyupgrade = "1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
